+++ src/app/(main)/profile/edit/_ui/ProfileEditContent.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { useState, useRef, useEffect, type ChangeEvent, type ComponentProps } from 'react'
+import { useState, useRef, type ChangeEvent, type ComponentProps } from 'react'
 import { useRouter } from 'next/navigation'
 import { useQuery } from '@tanstack/react-query'
 import { adopterQueries } from '@/entities/adopter'
@@ -11,6 +11,7 @@ import { useUpdateMyProfile } from '@/features/profile'
 import { useUploadSingleFile } from '@/features/upload'
 import { useLogout } from '@/features/auth'
 import { normalizeApiError } from '@/shared/api'
+import { useToast } from '@/shared/lib/useToast'
 import { WithdrawReason } from '@/shared/types'
 import {
   AlertMessage,
@@ -26,7 +27,6 @@ import {
 } from '@/shared/ui'
 import { AlertCircleIcon, CheckIcon } from '@/shared/assets/icons'
 
-const TOAST_DURATION_MS = 3000
 const NAME_MAX_LENGTH = 30
 const BIO_MAX_LENGTH = 200 // 서버 UpdateMyProfileRequestDto.bio maxLength
 
@@ -46,27 +46,6 @@ const LEAVE_DESCRIPTION = (
     복구되지 않습니다
   </>
 )
-
-// 토스트 표시 + 자동 닫힘 로직 분리 (SRP)
-// 성공(default)·실패(error) 두 종류를 같은 슬롯에서 표시한다.
-type ToastState = { message: string; status: 'default' | 'error' }
-
-const useToast = (duration = TOAST_DURATION_MS) => {
-  const [toast, setToast] = useState<ToastState | null>(null)
-
-  useEffect(() => {
-    if (!toast) return
-    const timer = setTimeout(() => setToast(null), duration)
-    return () => clearTimeout(timer)
-  }, [toast, duration])
-
-  return {
-    current: toast,
-    success: (message: string) => setToast({ message, status: 'default' }),
-    error: (message: string) => setToast({ message, status: 'error' }),
-    hide: () => setToast(null),
-  }
-}
 
 /** 프로필 편집 (Figma node 2145-191107) — GNB는 MainLayout 제공 */
 const ProfileEditContent = () => {
